% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{candyland}
\alias{candyland}
\title{Mock comprehensive data (CD) output file in a tidy data.frame}
\format{
A data frame of class \code{"tbl_df"} with 300 rows and 422 variables:
\describe{
  \item{crop_year}{year (2019-2021)}
}
}
\usage{
candyland
}
\description{
A dataset containing all variables of the original CD .xlsx file, and
representing 300 field-years.
}
\keyword{datasets}
