% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldprintr-package.R
\docType{package}
\name{fieldprintr-package}
\alias{fieldprintr}
\alias{fieldprintr-package}
\title{fieldprintr: Functions to complement the Fieldprint Platform}
\description{
Internal package for Field to Market that complements the Fieldprint Platform.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://field-to-market.github.io/fieldprintr/}
}

}
\author{
\strong{Maintainer}: Austin Pearce \email{apearce@fieldtomarket.org}

}
\keyword{internal}
