% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ftm_cd.R
\name{read_ftm_cd}
\alias{read_ftm_cd}
\title{read_ftm_cd}
\usage{
read_ftm_cd(file, sheet = "Report Data")
}
\arguments{
\item{file}{.xlsx comprehensive data output file as exported from Fieldprint}

\item{sheet}{the sheet of the .xlsx file to read (default = "Report Data")}
}
\value{
tidy data.frame (tibble)
}
\description{
This wrapper function standardizes the import of the Comprehensive Data (CD)
output file, a file that is exported from the Field to Market Fieldprint
Platform as an Excel .xlsx file.
}
\details{
The original name of an .xlsx CD file may be formatted like
"MM-DD-YYYY-####_Comprehensive_Data_My_Farm.xlsx". Regardless,
the read_ftm_cd() function looks for the sheet "Report Data", as exported
from the Fieldprint Platform.
}
\examples{
\dontrun{
raw_cd <- read_ftm_cd("data-raw/12-09-2022-1234_Comprehensive_Data_Candyland_Farms.xlsx")
}
}
