% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldprintr-package.R
\docType{package}
\name{fieldprintr-package}
\alias{fieldprintr}
\alias{fieldprintr-package}
\title{fieldprintr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://field-to-market.github.io/fieldprintr/}
}

}
\keyword{internal}
